Rework the carriage again
It’s much taller, theory being it gives the 2 bearings on the right more leverage and less likelihood to be detailed.
I coupled the mounts to the well constrained bearings on the right because I noticed that tensioning the carriage can cause the mounts to tilt, which I don’t want.
The bearing on the left is <variable> too far to the right and I’m using the print as a spring.